refactor(mobileapp): isolate useAuth to a dedicated file
